# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Transaction API.",
        "title": "SeaCloud APIs",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/files": {
            "get": {
                "description": "ListEntries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "ListEntries",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003csession_token\u003e",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Directory path",
                        "name": "dirpath",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ListEntriesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "UploadFiles",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "file"
                ],
                "summary": "UploadFiles",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003csession_token\u003e",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Directory path",
                        "name": "dirpath",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Files",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.UploadFileResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/directories": {
            "post": {
                "description": "CreateDirectory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "CreateDirectory",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003csession_token\u003e",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create directory request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateDirectoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/download": {
            "get": {
                "description": "DownloadFile",
                "tags": [
                    "file"
                ],
                "summary": "DownloadFile",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003csession_token\u003e",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File path",
                        "name": "file_path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/metadata": {
            "get": {
                "description": "GetMetadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "GetMetadata",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003csession_token\u003e",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File or directory path",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/file.Entry"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "file.Entry": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "full_path": {
                    "type": "string"
                },
                "is_dir": {
                    "type": "boolean"
                },
                "md5": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "mime_type": {
                    "type": "string"
                },
                "mode": {
                    "$ref": "#/definitions/os.FileMode"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateDirectoryRequest": {
            "type": "object",
            "required": [
                "dirpath"
            ],
            "properties": {
                "dirpath": {
                    "type": "string"
                }
            }
        },
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.ListEntriesResponse": {
            "type": "object",
            "properties": {
                "cursor": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/file.Entry"
                    }
                }
            }
        },
...
        "model.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "model.UploadFileResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "os.FileMode": {
            "type": "integer",
            "enum": [
                2147483648,
                1073741824,
                536870912,
                268435456,
                134217728,
                67108864,
                33554432,
                16777216,
                8388608,
                4194304,
                2097152,
                1048576,
                524288,
                2401763328,
                511,
                2147483648,
                1073741824,
                536870912,
                268435456,
                134217728,
                67108864,
                33554432,
                16777216,
                8388608,
                4194304,
                2097152,
                1048576,
                524288,
                2401763328,
                511
            ],
            "x-enum-comments": {
                "ModeAppend": "a: append-only",
                "ModeCharDevice": "c: Unix character device, when ModeDevice is set",
                "ModeDevice": "D: device file",
                "ModeDir": "d: is a directory",
                "ModeExclusive": "l: exclusive use",
                "ModeIrregular": "?: non-regular file; nothing else is known about this file",
                "ModeNamedPipe": "p: named pipe (FIFO)",
                "ModePerm": "Unix permission bits, 0o777",
                "ModeSetgid": "g: setgid",
                "ModeSetuid": "u: setuid",
                "ModeSocket": "S: Unix domain socket",
                "ModeSticky": "t: sticky",
                "ModeSymlink": "L: symbolic link",
                "ModeTemporary": "T: temporary file; Plan 9 only"
            },
            "x-enum-varnames": [
                "ModeDir",
                "ModeAppend",
                "ModeExclusive",
                "ModeTemporary",
                "ModeSymlink",
                "ModeDevice",
                "ModeNamedPipe",
                "ModeSocket",
                "ModeSetuid",
                "ModeSetgid",
                "ModeCharDevice",
                "ModeSticky",
                "ModeIrregular",
                "ModeType",
                "ModePerm"
            ]
        }
    }
}